requirements.txt
Django==6.0.4
django-cors-headers==4.7.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
Pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
drf-spectacular==0.29.0
django-cleanup==9.0.0
